requirements.txt
beautifulsoup4==4.12.2
pymongo==4.6.2
requests==2.31.0
apache-airflow-providers-mongo
